Bot/requirements.txt
discord.py
wavelink
python-dotenv